// src/app/data/interfaces/Respuestos.interface.ts
export interface Repuestos {
  capacitorArranque: boolean;
  capacitorPermanente: boolean;
  relay: boolean;
  contactor: boolean;
  controlador: boolean;
  sensor: boolean;
  capilar: boolean;
  fusibles: boolean;
  manillas: boolean;
  abanicoEvaporador: boolean;
  abanicoAxial: boolean;
  abanicoCondensador: boolean;
  pushBoton: boolean;
  tuboLedGrande: boolean;
  empaquePuerta: boolean;
  compresor: boolean;
  protectorTermico: boolean;
  tuboLedPequeno: boolean;
  interruptorLuz: boolean;
  resistenciaDrenaje: boolean;
  resistenciaEvaporador: boolean;
  valvulaCarga: boolean;
  filtroDeshidratador: boolean;
  resistenciaPuerta: boolean;
  vidrioPuertas: boolean;
}